requirements.txt

beautifulsoup4>=4.13.4
email-validator>=2.2.0
flask>=3.1.1
flask-dance>=7.1.0
flask-login>=0.6.3
flask-sqlalchemy>=3.1.1
gunicorn>=23.0.0
oauthlib>=3.2.2
openai>=1.79.0
pdfminer-six>=20250506
psycopg2-binary>=2.9.10
pyjwt>=2.10.1
requests>=2.32.3
trafilatura>=2.0.0
werkzeug>=3.1.3
email_validator
flask
flask-sqlalchemy
gunicorn
psycopg2-binary
